+++ accounts/managers.py
@@ -2,19 +2,23 @@ from django.contrib.auth.models import BaseUserManager
 
 
 class UserManager(BaseUserManager):
-    def create_user(self, email, full_name, password):
-        if not email:
-            raise ValueError('Email is required!')
-        if not full_name:
-            raise ValueError('full name is required!')
-
-        user = self.model(email=self.normalize_email(email), full_name=full_name)
+    def create_user(self, phone_number, password=None, is_manager=False, **extra_fields):
+        if not phone_number:
+            raise ValueError('The Phone Number must be set')
+        
+        user = self.model(phone_number=phone_number, is_manager=is_manager, **extra_fields)
         user.set_password(password)
-        user.save(using=self.db)
+
+        user.save(using=self._db)
         return user
 
-    def create_superuser(self, email, full_name, password):
-        user = self.create_user(email, full_name, password)
-        user.is_admin = True
-        user.save(using=self.db)
-        return user
+    def create_superuser(self, phone_number, password=None, **extra_fields):
+        extra_fields.setdefault('is_admin', True)
+        extra_fields.setdefault('is_active', True)
+        extra_fields.setdefault('is_verify', True)
+        extra_fields.setdefault('is_manager', True)  # Optionally make superusers managers too
+    
+        if extra_fields.get('is_admin') is not True:
+            raise ValueError('Superuser must have is_admin=True.')
+        
+        return self.create_user(phone_number, password, **extra_fields)
