+++ accounts/models.py
@@ -3,30 +3,6 @@ from django.contrib.auth.models import AbstractBaseUser
 
 from phonenumber_field.modelfields import PhoneNumberField
 from .managers import UserManager
-#from shop.models import Product
-from django.contrib.auth.models import BaseUserManager
-
-class UserManager(BaseUserManager):
-    def create_user(self, phone_number, password=None, is_manager=False, **extra_fields):
-        if not phone_number:
-            raise ValueError('The Phone Number must be set')
-        
-        user = self.model(phone_number=phone_number,is_manager=is_manager, **extra_fields)
-        user.set_password(password)
-
-        user.save(using=self._db)
-        return user
-
-    def create_superuser(self, phone_number, password=None, **extra_fields):
-        extra_fields.setdefault('is_admin', True)
-        extra_fields.setdefault('is_active', True)
-        extra_fields.setdefault('is_verify', True)
-        extra_fields.setdefault('is_manager', True)  # Optionally make superusers managers too
-    
-        if extra_fields.get('is_admin') is not True:
-            raise ValueError('Superuser must have is_admin=True.')
-        
-        return self.create_user(phone_number, password, **extra_fields)
 
 class User(AbstractBaseUser):
     phone_number = PhoneNumberField(unique=True, null=True, blank=True)
@@ -35,7 +11,7 @@ class User(AbstractBaseUser):
     is_admin = models.BooleanField(default=False)
     is_active = models.BooleanField(default=False)
     is_verify = models.BooleanField(default=False)
-    #likes = models.ManyToManyField(Product, blank=True, related_name='likes')
+    likes = models.ManyToManyField('shop.Product', blank=True, related_name='likes')
     # set a manager role for shop manager to access orders and products
     is_manager = models.BooleanField(default=False)
     created_at = models.DateTimeField(auto_created=True,auto_now_add=True)
