+++ dashboard/templates/add_category.html
@@ -5,7 +5,7 @@
     {% csrf_token %}
     {{ form }}
     <div class="form-group">
-        <button class="btn btn-primary mt-3 w-100" type="submit">Submit</button>
+        <button class="btn btn-primary mt-3 w-100" type="submit">ثبت</button>    
     </div>
 </form>
