+++ dashboard/urls.py
@@ -1,4 +1,4 @@
-from django.urls import path
+from django.urls import path, include,re_path
 
 from dashboard import views
