+++ orders/templates/checkout.html
@@ -4,11 +4,11 @@
 <div class="col-md-2"></div>
 
 <div class="col-md-8 border rounded p-3">
-    <h3 class="text-muted">Your Order Amount is: ${{order.get_total_price}}</h3>
+    <h3 class="text-muted">مبلغ سفارش شما: ${{order.get_total_price}}</h3>
     <hr>
-    <p><b>Do you want to continue and pay?</b></p>
-    <a href="{% url 'cart:show_cart' %}" class="btn btn-danger">Back To Cart</a>
-    <a href="{% url 'orders:create_order' %}" class="btn btn-outline-success">Continue and pay</a>
+    <p><b>آیا مایل به ادامه فرآیند هستید؟</b></p>
+    <a href="{% url 'cart:show_cart' %}" class="btn btn-danger">بازگشت به سبد خرید</a>
+    <a href="{% url 'orders:create_order' %}" class="btn btn-outline-success">ادامه فرآیند و پرداخت</a>
 <div class="col-md-2"></div>
 
 {% endblock %}
