+++ shop/templates/favorites.html
@@ -1,7 +1,7 @@
 {% extends 'base.html' %}
 {% block content %}
 {% if products %}
-<h2>Your Favorites</h2>
+<h2>علاقه مندی ها</h2>
 <hr>
 {% for product in products %}
 <div class="card me-2 mb-2" style="width: 16rem;">
@@ -9,9 +9,9 @@
   <div class="mt-3 text-center">
     <h5 class="card-title">{{ product.title }}</h5>
     <p class="text-muted">${{ product.price }}</p>
-    <a href="{{ product.get_absolute_url }}" class="mb-3 btn btn-primary w-100">Buy Now</a>
+    <a href="{{ product.get_absolute_url }}" class="mb-3 btn btn-primary w-100">خرید</a>
   </div>
-  <a href="{% url 'shop:remove_from_favorites' product.id %}" class="mb-3 btn btn-outline-danger">Remove</a>
+  <a href="{% url 'shop:remove_from_favorites' product.id %}" class="mb-3 btn btn-outline-danger">حذف</a>
 </div>
 {% endfor %}
 {% else %}
@@ -19,8 +19,8 @@
   <div class="col-md-2"></div>
   <div class="col-md-8 mt-5 pt-5 text-center">
     <br>
-    <h3 class="text-muted text-capitalize">favorites list is empty..!</h3>
-    <a href="/" class="text-primary text-decoration-none">Back</a>
+    <h3 class="text-muted text-capitalize">لیست علاقه مندی ها خالی است..!</h3>
+    <a href="/" class="text-primary text-decoration-none">بازگشت</a>
   </div>
   <div class="col-md-2"></div>
 </div>
