+++ shop/templates/home_page.html
@@ -2,22 +2,31 @@
 
 {% block content %}
 {% if products %}
+<div class="row">
 {% for product in products.object_list %}
-<div class="card me-2 mb-2" style="width: 16rem;">
-    <img style="object-fit: cover;" class="card-img mt-2" width="268" height="200" src="{{ product.image.url }}">
-    <div class="mt-3 text-center">
+<div class="col-md-3 mb-4">
+  <div class="card h-100 shadow-sm">
+    <img style="object-fit: cover;" class="card-img-top" height="200" src="{{ product.image.url }}" alt="{{ product.title }}">
+    <div class="card-body">
       <h5 class="card-title">{{ product.title }}</h5>
-      <p class="text-muted">${{ product.price }}</p>
-      <a href="{{ product.get_absolute_url }}" class="mb-3 btn btn-outline-primary w-100">Buy Now</a>
+      
+      
+      <div class="d-flex justify-content-between align-items-center mt-3">
+        <h6 class="m-0 text-success">${{ product.price }}</h6>
+        <a href="{{ product.get_absolute_url }}" class="btn btn-sm btn-outline-primary">مشاهده جزئیات</a>
+      </div>
     </div>
   </div>
+</div>
 {% endfor %}
+</div>
+
 <!-- pagination -->
 <center class="mt-5">
   <div class="col-md-2">
     <ul class="pagination">
       {% if products.has_previous %}
-      <li class="page-item"><a class="page-link" href="?page={{ products.previous_page_number }}">Previous</a></li>
+      <li class="page-item"><a class="page-link" href="?page={{ products.previous_page_number }}">بازگشت</a></li>
       <li class="page-item"><a class="page-link" href="?page={{ products.previous_page_number }}">{{products.previous_page_number}}</a></li>
       {% endif %}
       <li class="page-item"><a class="page-link" href="?page={{ products.number }}">{{products.number}}</a></li>
@@ -33,7 +42,7 @@
   <div class="col-md-2"></div>
   <div class="col-md-8 mt-5 pt-5 text-center">
     <br>
-    <h3 class="text-muted text-capitalize">No Products..!</h3>
+    <h3 class="text-muted text-capitalize">محصولی یافت نشد..!</h3>
   </div>
   <div class="col-md-2"></div>
 </div>
