+++ shop/templates/product_detail.html
@@ -8,25 +8,25 @@
     <div class="col-md-5 p-3 border rounded">
         <!-- category -->
         <div class="border-bottom mb-2 pb-2">
-            <span class="text-muted ">categories:</span><a class="text-decoration-none text-dark" href="{% url 'shop:filter_by_category' product.category.slug %}"> {{product.category}}</a>
+            <span class="text-muted ">دسته بندی:</span><a class="text-decoration-none text-dark" href="{% url 'shop:filter_by_category' product.category.slug %}"> {{product.category}}</a>
         </div>
         <!-- title -->
         <h2>{{ product.title }}</h2>
         <!-- price -->
-        <h4 class="mt-4 text-dark">Price: ${{ product.price }}</h4>
+        <h4 class="mt-4 text-dark">قیمت: ${{ product.price }}</h4>
         <!-- description -->
         <div class="mt-4 pe-3 mb-5">{{ product.description }}</div>
         <!-- cart btn -->
         <form method="post" action="{% url 'cart:add_to_cart' product.id %}">
             {% csrf_token %}
             {{ form }}
-            <input type="submit" class="btn btn-primary mt-4" value="Add to cart">
+            <input type="submit" class="btn btn-primary mt-4" value="افزودن به سبد خرید">
         </form>
         <br>
         {% if favorites == 'remove' %}
-        <a href="{% url 'shop:remove_from_favorites' product.id %}" class="btn btn-outline-danger mt-5 text-capitalize ">{{favorites}} from favorites</a>
+        <a href="{% url 'shop:remove_from_favorites' product.id %}" class="btn btn-outline-danger mt-5 text-capitalize ">حذف از علاقه مندی ها</a>
         {% else %}
-        <a href="{% url 'shop:add_to_favorites' product.id %}" class="btn btn-outline-success mt-5 text-capitalize ">add to {{favorites}}</a>
+        <a href="{% url 'shop:add_to_favorites' product.id %}" class="btn btn-outline-success mt-5 text-capitalize ">افزودن به علاقه مندی ها</a>
         {% endif %}
     </div>
     <!-- product image -->
@@ -36,7 +36,7 @@
 </div>
 <!-- related products -->
 <div class="row mt-4 mb-4">
-    <h3>Related Products:</h3><hr>
+    <h3>محصولات مرتبط:</h3><hr>
     {% for p in related_products %}
         <!-- dont show the current product in this page -->
         {% if p != product %}
@@ -45,7 +45,7 @@
                 <div class="mt-3 text-center">
                 <h5 class="card-title">{{ p.title }}</h5>
                 <p class="text-muted">${{ p.price }}</p>
-                <a href="{{ p.get_absolute_url }}" class="mb-3 btn btn-outline-primary w-100">Buy Now</a>
+                <a href="{{ p.get_absolute_url }}" class="mb-3 btn btn-outline-primary w-100">خرید</a>
                 </div>
             </div>
         {% endif %}
